requirements.txt
faiss-cpu==1.8.0
numpy==1.26.4
PyPDF2==3.0.1
rouge-score==0.1.2
sentence-transformers==2.7.0
tqdm==4.66.4
